+++ invenio_vocabularies/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2020-2023 CERN.
+# Copyright (C) 2020-2024 CERN.
 #
 # Invenio-Vocabularies is free software; you can redistribute it and/or
 # modify it under the terms of the MIT License; see LICENSE file for more
@@ -10,6 +10,6 @@
 
 from .ext import InvenioVocabularies
 
-__version__ = "2.4.0"
+__version__ = "3.2.0"
 
 __all__ = ("__version__", "InvenioVocabularies")