+++ invenio_vocabularies/assets/semantic-ui/js/invenio_vocabularies/src/contrib/forms/Funding/CustomAwardForm.js
@@ -18,13 +18,18 @@ function CustomAwardForm({ deserializeFunder, selectedFunding }) {
   function deserializeFunderToDropdown(funderItem) {
     let funderName = null;
     let funderPID = null;
+    let funderCountry = null;
 
     if (funderItem.name) {
       funderName = funderItem.name;
     }
 
-    if (funderItem.pid) {
-      funderPID = funderItem.pid;
+    if (funderItem.id) {
+      funderPID = funderItem.id;
+    }
+
+    if (funderItem.country) {
+      funderCountry = funderItem.country;
     }
 
     if (!funderName && !funderPID) {
@@ -32,7 +37,7 @@ function CustomAwardForm({ deserializeFunder, selectedFunding }) {
     }
 
     return {
-      text: funderName || funderPID,
+      text: [funderName, funderCountry, funderPID].filter((val) => val).join(", "),
       value: funderItem.id,
       key: funderItem.id,
       ...(funderName && { name: funderName }),