+++ invenio_vocabularies/records/systemfields/relations.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2022 CERN.
+# Copyright (C) 2022-2024 CERN.
 #
 # Invenio-Records-Resources is free software; you can redistribute it and/or
 # modify it under the terms of the MIT License; see LICENSE file for more
@@ -38,7 +38,7 @@ class CustomFieldsRelation(RelationsField):
                 relations[cf.name] = cf.relation_cls(
                     f"custom_fields.{cf.name}",
                     keys=cf.field_keys,
-                    pid_field=Vocabulary.pid.with_type_ctx(cf.vocabulary_id),
+                    pid_field=cf.pid_field,
                     cache_key=cf.vocabulary_id,
                 )